XML Pull Parser optimizations
This patch includes three improvements to the pull parser.

1. readValue now only uses the stringPool for all-whitespace strings.
Values are generally not repeated, and this function can
thrash the string pool in some cases

2. Added a fast path for the normal, correct, end-tag case.
This avoids a hash and lookup in the stringPool

3. While parsing an element (not attribute) name, the namespace and
post-namespace strings are generated using the stringPool.
This saves the cost of calling String.substring in adjustNsp.

Change-Id: Iea6c6e42d9f24252ca63d61bcf06dc02d23ddad3